src/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.4
opencv-python-headless>=4.10.0
torch>=2.3.0
torchvision>=0.18.0
scikit-learn>=1.5.0
pytest>=8.2.1
requests>=2.32.2
